moved the gamelogic to a separete JS file for better readability
